requirements.txt
# Core scraping dependencies
requests>=2.31.0
PyYAML>=6.0.1
pandas>=2.0.0
openpyxl>=3.1.0
playwright>=1.40.0
beautifulsoup4>=4.12.0

# Streamlit web app
streamlit>=1.28.0

# Optional dependencies for scheduling
schedule>=1.2.0
pytz>=2023.3
